deps/v8/test/webkit/fast/js/Promise-static-resolve-expected.txt

Test Promise.resolve

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS result is undefined
PASS result is "hello"
PASS successfullyParsed is true

TEST COMPLETE
